made auth owner of roles model
