--- src/models.ts
@@ -1,10 +1,0 @@
-export enum Role {
-	ADMIN = "Admin",
-	STAFF = "Staff",
-	MENTOR = "Mentor",
-	APPLICANT = "Applicant",
-	ATTENDEE = "Attendee",
-	USER = "User",
-	SPONSOR = "Sponsor",
-	BLOBSTORE = "Blobstore",
-}
