+++ src/services/auth/auth-models.ts
@@ -1,5 +1,3 @@
-import { Role } from "../../models.js";
-
 export interface ProfileData {
 	id?: string,
 	email: string,
@@ -11,6 +9,17 @@ export interface JwtPayload {
 	provider: string,
 	roles: Role[],
 	exp?: number
+}
+
+export enum Role {
+	ADMIN = "Admin",
+	STAFF = "Staff",
+	MENTOR = "Mentor",
+	APPLICANT = "Applicant",
+	ATTENDEE = "Attendee",
+	USER = "User",
+	SPONSOR = "Sponsor",
+	BLOBSTORE = "Blobstore",
 }
 
 export enum Provider {
